requirements.txt
# Trading System - Python Dependencies
# Compatible with Mac M4 (ARM64)

# Core Data & Analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Market Data APIs
yfinance==0.2.33
nsepy==0.8

# Database
psycopg2-binary==2.9.9  # TimescaleDB/PostgreSQL
sqlalchemy==2.0.23

# Caching & Messaging
redis==5.0.1
kafka-python==2.0.2

# Scheduling & Background Tasks
schedule==1.2.0
apscheduler==3.10.4

# Web Dashboard
dash==2.14.2
plotly==5.18.0
flask==3.0.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0

# Optional: Performance
# uvloop==0.19.0  # Faster event loop (for async if needed)
# orjson==3.9.10  # Faster JSON (if needed)
